Move arbitrage opportunity functions to ArbitrageService
Refactor arbitrage functions into smaller functions
Arbitrage functions code cleanup
